♻️ [Manager Page] Member/MemberBox Type 분리
