+++ src/types/IMemberValue.ts
@@ -1,0 +1,19 @@
+import { Dispatch, SetStateAction } from "react";
+import { IMemberBoxValue } from "./IMemberBoxValue";
+
+export interface IMemberValue {
+  name: string;
+  schoolNumber: string;
+  phoneNumber: string;
+  introduce: string;
+}
+export interface IMemberPropsValue {
+  state: IMemberBoxValue[];
+  setState: Dispatch<SetStateAction<IMemberBoxValue[]>>;
+  userIndex: number;
+  BeforeContainerIndex: number;
+  name: string;
+  schoolNumber: string;
+  phoneNumber: string;
+  introduce: string;
+}
